Preview is frontend zone, fix for viewing content in another locale
